RavenDB-24962 - use a single buffer per transaction
