+++ src/Voron/Impl/Transaction.cs
@@ -24,6 +24,19 @@ namespace Voron.Impl
         private Dictionary<Tuple<Tree, Slice>, Tree> _multiValueTrees;
         private Dictionary<long, ByteString> _cachedDecompressedBuffersByStorageId;
 
+        private StreamBufferAllocator.Buffer _streamBuffer;
+
+        public StreamBufferAllocator.Buffer StreamBuffer
+        {
+            get
+            {
+                if (_streamBuffer != null)
+                    return _streamBuffer;
+
+                return _streamBuffer = StreamBufferAllocator.Instance.Rent();
+            }
+        }
+
         internal Dictionary<long, ByteString> CachedDecompressedBuffersByStorageId =>
             _cachedDecompressedBuffersByStorageId ??= new Dictionary<long, ByteString>();
 
@@ -534,8 +547,10 @@ namespace Voron.Impl
                 }
             }
 
-            _lowLevelTransaction?.Dispose();
+            _streamBuffer?.Dispose();
+            _streamBuffer = null;
 
+            _lowLevelTransaction?.Dispose();
             _lowLevelTransaction = null;
         }
